requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-wtf==1.1.1
flask-migrate==4.0.4
email-validator==2.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
paypalrestsdk==1.13.1
requests==2.31.0
pillow==10.0.0
flask-mail==0.9.1
python-slugify==8.0.1
flask-ckeditor==0.4.6
flask-cors==4.0.0
gunicorn==21.2.0